requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
streamlit>=1.10.0
pytest>=7.0.0
numpy-stl>=2.16.0
plotly>=5.5.0 